refactor[#4] : 패키지 이동
- 기존에 패키지 경로가 프로젝트 전체 구성과 맞지 않았기 때문에, 일관성 유지를 위하여 옮겨주었다.
